Renamed the SearchBarAutoComplete.
The SearchBarAutoComplete now can display user information in a modal.
